requirements.txt
selenium>=4.19.0
undetected-chromedriver>=3.5.5
PyYaml>=6.0.1